core/views/auth_views.py
"""
Authentication views for EduPulse project.

This module contains views for user authentication and management.
Dev 1 responsibility: Auth & User Management
"""

from rest_framework import status
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import AllowAny, IsAuthenticated
from rest_framework.response import Response
from rest_framework.views import APIView
from rest_framework_simplejwt.tokens import RefreshToken
from django.contrib.auth import authenticate
from django.contrib.auth.models import User
from ..serializers import (
    RegisterSerializer, LoginSerializer, UserSerializer, UserProfileSerializer,
    PasswordResetRequestSerializer, PasswordResetConfirmSerializer, ChangePasswordSerializer
)


class RegisterView(APIView):
    """
    API view for user registration.
    """
    permission_classes = [AllowAny]
    
    def post(self, request):
        """
        Register a new user.
        """
        # TODO: Implement user registration logic
        # 1. Validate registration data
        # 2. Create user and profile
        # 3. Generate JWT tokens
        # 4. Return user data and tokens
        return Response({
            'message': 'User registration endpoint - Dev 1 to implement',
            'status': 'success'
        }, status=status.HTTP_201_CREATED)


class LoginView(APIView):
    """
    API view for user login.
    """
    permission_classes = [AllowAny]
    
    def post(self, request):
        """
        Authenticate user and return JWT tokens.
        """
        # TODO: Implement user login logic
        # 1. Validate login credentials
        # 2. Authenticate user
        # 3. Generate JWT tokens
        # 4. Track login session
        # 5. Return tokens and user data
        return Response({
            'message': 'User login endpoint - Dev 1 to implement',
            'status': 'success'
        }, status=status.HTTP_200_OK)


class LogoutView(APIView):
    """
    API view for user logout.
    """
    permission_classes = [IsAuthenticated]
    
    def post(self, request):
        """
        Logout user and invalidate tokens.
        """
        # TODO: Implement user logout logic
        # 1. Invalidate JWT tokens
        # 2. Update session status
        # 3. Clear any cached data
        return Response({
            'message': 'User logout endpoint - Dev 1 to implement',
            'status': 'success'
        }, status=status.HTTP_200_OK)


class UserProfileView(APIView):
    """
    API view for user profile management.
    """
    permission_classes = [IsAuthenticated]
    
    def get(self, request):
        """
        Get current user's profile.
        """
        # TODO: Implement profile retrieval logic
        # Return user profile data
        return Response({
            'message': 'Get user profile endpoint - Dev 1 to implement',
            'status': 'success'
        }, status=status.HTTP_200_OK)
    
    def put(self, request):
        """
        Update current user's profile.
        """
        # TODO: Implement profile update logic
        # Update user profile data
        return Response({
            'message': 'Update user profile endpoint - Dev 1 to implement',
            'status': 'success'
        }, status=status.HTTP_200_OK)


class PasswordResetRequestView(APIView):
    """
    API view for password reset request.
    """
    permission_classes = [AllowAny]
    
    def post(self, request):
        """
        Request password reset token.
        """
        # TODO: Implement password reset request logic
        # 1. Validate email
        # 2. Generate reset token
        # 3. Send email with reset link
        return Response({
            'message': 'Password reset request endpoint - Dev 1 to implement',
            'status': 'success'
        }, status=status.HTTP_200_OK)


class PasswordResetConfirmView(APIView):
    """
    API view for password reset confirmation.
    """
    permission_classes = [AllowAny]
    
    def post(self, request):
        """
        Confirm password reset with token.
        """
        # TODO: Implement password reset confirmation logic
        # 1. Validate token
        # 2. Update password
        # 3. Invalidate token
        return Response({
            'message': 'Password reset confirmation endpoint - Dev 1 to implement',
            'status': 'success'
        }, status=status.HTTP_200_OK)


class ChangePasswordView(APIView):
    """
    API view for changing user password.
    """
    permission_classes = [IsAuthenticated]
    
    def post(self, request):
        """
        Change user password.
        """
        # TODO: Implement password change logic
        # 1. Validate old password
        # 2. Update to new password
        # 3. Invalidate existing sessions
        return Response({
            'message': 'Change password endpoint - Dev 1 to implement',
            'status': 'success'
        }, status=status.HTTP_200_OK)


class RefreshTokenView(APIView):
    """
    API view for refreshing JWT tokens.
    """
    permission_classes = [AllowAny]
    
    def post(self, request):
        """
        Refresh JWT access token.
        """
        # TODO: Implement token refresh logic
        # 1. Validate refresh token
        # 2. Generate new access token
        # 3. Return new tokens
        return Response({
            'message': 'Token refresh endpoint - Dev 1 to implement',
            'status': 'success'
        }, status=status.HTTP_200_OK) 